refactor: adds new layer for domains
