+++ src/core/products/entities/Product.js
