+++ src/core/products/exceptions/InvalidCategoryError.js
